+++ lib/project_types/node/commands/connect.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 module Node
   class Command
-    class Connect < ShopifyCLI::SubCommand
+    class Connect < ShopifyCLI::Command::AppSubCommand
       prerequisite_task ensure_project_type: :node
 
       def call(*)