+++ lib/project_types/node/commands/create.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 module Node
   class Command
-    class Create < ShopifyCLI::SubCommand
+    class Create < ShopifyCLI::Command::AppSubCommand
       prerequisite_task :ensure_authenticated
 
       options do |parser, flags|