+++ test/project_types/php/commands/open_test.rb
@@ -13,7 +13,7 @@ module PHP
       def test_run
         ShopifyCLI::Context.any_instance.expects(:open_url!)
           .with("https://example.com/login?shop=my-test-shop.myshopify.com")
-        run_cmd("php open")
+        run_cmd("app php open")
       end
     end
   end